docs(f45): as quatro atas, o plano, o relatorio e os indices
Quatro atas em docs/DECISOES.md — as tres que a ordem exige (a ata do portao com
a consequencia mecanica e o bypass; onde moram os .test.tsx e o que aconteceu com
as cinco varreduras repo-wide; o formato da linha FIM e por que universo vazio
levanta excecao) mais uma quarta, que registra o que esta fase NAO conseguiu
provar e por que se mergeou assim mesmo.

docs/PLAN-F45.md e docs/RELATORIO-F45.md. O relatorio traz, no topo, as duas
pendencias — o portao nao ligado (sem gh) e o SQL nunca executado (sem Docker) —
e o §8 tem o comando gh api pronto para colar mais o passo a passo do painel.

README.md registra a decisao do portao com data, como o §5/Risco da ficha manda,
inclusive o risco de ela ser configuracao FORA do repositorio que nenhuma trava
interna defende.

Correcao de revisao: a primeira versao da trava fazia import('../../vitest.config.mts'),
que o tsc recusa (TS5097) e nem o lint nem o build enxergam. Passou a ler a config
como TEXTO, como ja fazia com o YAML — afrouxar o tsconfig inteiro para um teste
ficar verde e a troca que esta fase existe para nao fazer.

+++ supabase/tests/asserts_ferramenta.sql
@@ -52,7 +52,7 @@ begin
   -- 1b. universo NULL é o mesmo caso (um `select count(*) into` que não achou
   --     linha devolve null, e é assim que a tautologia costuma chegar).
   begin
-    v_res := pg_temp.assert_zero_de('autoteste 1b', 0, null);
+    v_res := pg_temp.assert_zero_de('autoteste 1b', 0, null::bigint);
     v_falhas := v_falhas + 1;
     raise warning '✗ 1b assert_zero_de ACEITOU universo NULL (devolveu %)', v_res;
   exception when others then
@@ -102,7 +102,7 @@ begin
 
   -- 4b. contagem de ruins NULA idem.
   begin
-    v_res := pg_temp.assert_zero_de('autoteste 4b', null, 7);
+    v_res := pg_temp.assert_zero_de('autoteste 4b', null::bigint, 7);
     v_falhas := v_falhas + 1;
     raise warning '✗ 4b aceitou contagem de ruins NULA (devolveu %)', v_res;
   exception when others then
